add expungeeverything
